Move TimeRange component to shared folders
