+++ app/shared/time-range/index.js
@@ -1,0 +1,3 @@
+import TimeRange from './TimeRange';
+
+export default TimeRange;
